move add and delete functionality to playerWallet handler
